refactor(client/src): add utils folder
